requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
motor==3.3.2
pymongo==4.6.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
cryptography==42.0.0
python-multipart==0.0.6

# Email & SMTP
aiosmtplib==3.0.1

# OAuth & Google Integration
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Async Utilities
httpx==0.26.0
aiohttp==3.9.3

# Logging & Monitoring
python-json-logger==2.0.7

# Rate Limiting
slowapi==0.1.9

# Environment Variables
python-dotenv==1.0.0

# Additional Utilities
tenacity==8.2.3
python-dateutil==2.8.2
pytz==2024.1